feat(category): 添加分类管理功能- 在分类页面添加父分类选择功能 - 实现分类列表按父分类筛选 - 添加分类名称递归查询方法 - 新增获取分类及其子分类ID列表的方法 - 更新产品列表和详情页面优化产品快照获取逻辑，显示完整分类名称

+++ mall-core/src/main/java/com/jiang/mall/service/IProductService.java
@@ -39,7 +39,7 @@ public interface IProductService extends IService<Product> {
 
     Boolean deleteProduct(Long id);
 
-    Integer queryStoksById(Long productId);
+    Long queryStoksById(Long productId);
 
     Boolean queryCode(String code);
